Initial import of LAR algorithm.
This is done by binding minilearn library. It is still a work in
progress.

As cblas is needed by minilearn, a stripped down version of cblas
is shipped in case distutils does not find a suitable cblas
implementation to link with.

In linux systems, having atlas installed will be more than enough
to link against a decent cblas.