Add a reinitialization function for estimators + write test for GridSearch.
We wish pipelines and estimators can be used in the same, transparent
way. This was not the case. For instance, in the GridSearch module,
where the estimator under optimization need to be reconstructed with a
new set of parameters, the use of a pipeline as the estimator to
optimize was impossible. Indeed, the pipeline reconstruction was
impossible because only we only had *references* to its constituing
estimators.
To make the reconstruction of any kind of estimator (pipeline or
simple estimator), we added a _reinit() method in the BaseEstimator
class. This method perform a deep copy and a reinitialization of an
estimator object, without modifying the original one. We applied this
modification in the grid_search module and added a test for the
GridSearch function to garanty its behavior.
Note that there is no need to reimplement the _reinit function in case
one need to construct his own estimator : the purpose of this function
is to be overwritten in the Pipeline class, for the reasons we
mentionned above.