Since we changed the name of the figure names, changing the rst files.
In a previous commit, I changed the way the figures files where
generated. Each figure file is now similar to plot_example_<n>.png,
where n is the number of the figure. Therefore, even if there is only
one figure generated, the name of the file will be plot_example_1.png.