More covariance refactoring: separate MLE computation from object.
It had to be possible to fit a simple Maximum Likelihood Estimator of
covariance without having to instantiate a new Covariance object, as
is can result in a loss in code clarity and in performances.  The
separation of algorithm and object follows the model used in
shrun_covariance_.py classes.